% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{all.results.class-class}
\alias{all.results.class}
\alias{all.results.class-class}
\title{all.results.class}
\description{
all.results.class
}
\section{Slots}{

\describe{
\item{\code{attribute.profile.result}}{An object of \code{\link{attribute.profile.class}}}

\item{\code{attribute.result}}{An object of \code{\link{attribute.class}}}

\item{\code{parameter.result}}{An object of \code{\link{parameter.class}}}
}}

