% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{attribute.class-class}
\alias{attribute.class}
\alias{attribute.class-class}
\title{attribute.class}
\description{
attribute.class
}
\section{Slots}{

\describe{
\item{\code{results}}{A data frame of probability of mastering each attribute}
}}

