% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{attribute.profile.class-class}
\alias{attribute.profile.class}
\alias{attribute.profile.class-class}
\title{attribute.profile.class}
\description{
attribute.profile.class
}
\section{Slots}{

\describe{
\item{\code{results}}{A data frame of probability of mastering each attribute profile}

\item{\code{attribute.profile.matrix}}{A matrix of all possible attribute profiles}
}}

