% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CalculatePiIStarPrime}
\alias{CalculatePiIStarPrime}
\title{Calculate value for pi.i.star.prime for one item}
\usage{
CalculatePiIStarPrime(pi.i.star, r.ia.star, q.ia)
}
\arguments{
\item{pi.i.star}{a numeric value for the probability that a person responds correctly given they have not mastered
any attributes before being penalized for not mastering an attribute.}

\item{r.ia.star}{a numeric 1 x nattributes vector of the penalty probabilities for this item of a correct response
for each required attribute that has not been mastered.}

\item{q.ia}{a numeric 1 x nattributes vector of the qmatrix entries for the given item.}
}
\value{
a numeric value representing the pi.i.star_prime for item i
}
\description{
Given an NCRUM parameterization and values for pi.i.star, r.ia.star, and qmatrix, calculates the value for
for pi.i.star.prime for the given item
}

