% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{class.probabilities.test}
\alias{class.probabilities.test}
\title{Class probabilities for Q-matrix containing no interaction between attributes}
\format{a data frame containing 15 students and 3 attributes. 1 student per row}
\usage{
class.probabilities.test
}
\description{
This data set list the class probabities for Q-matrix containing no interaction between attributes.
This data set is used to check output of \code{ScoreDCM}
}
\author{
Margi Dubal \email{margidubal@gmail.com} & Diane Losardo \email{dlosardo@amplify.com}
}

