% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{dcm.scorer.class-class}
\alias{dcm.scorer.class}
\alias{dcm.scorer.class-class}
\title{dcm.scorer.class}
\description{
dcm.scorer.class
}
\section{Slots}{

\describe{
\item{\code{inputs}}{A list of inputs provided by user}

\item{\code{mcmc.inputs}}{A list of inputs for MCMC estimation}

\item{\code{results}}{An object of \code{\link{all.results.class}}}

\item{\code{mcmc.output}}{A list of MCMC runs}
}}

