% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DrawAlphas}
\alias{DrawAlphas}
\title{Draw values from multinomial distribution}
\usage{
DrawAlphas(pmatrix, y, nus0)
}
\arguments{
\item{pmatrix}{a matrix indicating all possible attribute profiles}

\item{y}{number of random vectors to draw}

\item{nus0}{a numeric vector of length \code{nclasses} for nu parameter estimates}
}
\value{
a numeric vector of alpha values
}
\description{
Draws values from a mutinomial distribution and matches them with a row in the pmatrix (attribute profile matrix)
to obtain the alpha values
}

