% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DrawClasses}
\alias{DrawClasses}
\title{Draw latent variable values}
\usage{
DrawClasses(nattributes, class0, nus0, observations, nobservations,
  threshold.values, pmatrix)
}
\arguments{
\item{nattributes}{a numeric value indicating number of alphas to be drawn}

\item{class0}{a numeric vector of previous classes}

\item{nus0}{a numeric vector of length \code{nclasses} for nu parameter estimates}

\item{observations}{a data frame or matrix of dichotomous responses}

\item{nobservations}{a numeric value indicating number of rows of the observation data frame or matrix}

\item{threshold.values}{an nclasses by nitems numeric matrix with appropriate item threshold values}

\item{pmatrix}{a numeric nclasses by nattributes matrix of all possible attribute profiles}
}
\value{
a numeric vector of classes indicating one of the possible attribute profile for each student
}
\description{
Produces values of alphas(s) (i.e., latent variable(s)) by using a rejection/acceptance Metropolis-Hastings procedure.
}

