% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetAllProbsCorrectNcrum}
\alias{GetAllProbsCorrectNcrum}
\title{Calculate all item probabilities of correct response for NCRUM parameterization}
\usage{
GetAllProbsCorrectNcrum(pi.i.star.prime, r.ia.star.qmatrix, qmatrix, alphas)
}
\arguments{
\item{pi.i.star.prime}{a numeric 1 x nitems vector of the probabilities of a correct response for each item given
that no attributes have been mastered.}

\item{r.ia.star.qmatrix}{a numeric nitems x nattributes vector of the penalty probabilities for each item of
a correct response for each required attribute that has not been mastered.}

\item{qmatrix}{a numeric nitems by nattributes matrix that specifies which items are required for mastery of each attribute (i.e., latent variable)}

\item{alphas}{a numeric 1 x n_attributes vector with a 1 if a student has mastered an attribute and 0 otherwise.}
}
\value{
probabilities of a correct response for all items given a person's attribute profile.
}
\description{
Calculates the probabilities of a correct response for all items given an NCRUM parameterization
}

