% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetAttributeProfiles}
\alias{GetAttributeProfiles}
\title{Attribute Profiles}
\usage{
GetAttributeProfiles(nattributes, attribute.names = NULL, profile.names = NULL)
}
\arguments{
\item{nattributes}{a numeric value for number of attributes}

\item{attribute.names}{optional character vector of attribute names}

\item{profile.names}{optional character vector of attribute profile names}
}
\value{
a matrix of dimension \code{2^nattributes X nattributes} containing binary values
}
\description{
Given a number of attributes, it generates all possible attribute profiles (response patterns of latent variables)
}
\examples{
\dontrun{

 GetAttributeProfiles(nattributes=3)
}
}
\author{
Margi Dubal \email{margidubal@gmail.com} & Diane Losardo \email{dlosardo@amplify.com}
}

