% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetAttributesProbabilitiesMCMC}
\alias{GetAttributesProbabilitiesMCMC}
\title{Calculate attribute probabilities}
\usage{
GetAttributesProbabilitiesMCMC(class.result, pmatrix)
}
\arguments{
\item{class.result}{a list of classes containing all interactions and chains of MCMC}

\item{pmatrix}{a numeric nclasses by nattributes matrix of all possible attribute profiles}
}
\description{
Calculates the attribute (i.e., latent variable) probabilities for all students
}

