% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetClassProbabilitiesMCMC}
\alias{GetClassProbabilitiesMCMC}
\title{Calculate attribute profile probabilities}
\usage{
GetClassProbabilitiesMCMC(class.result, nclasses)
}
\arguments{
\item{class.result}{a list of classes containing all interactions and chains of MCMC}

\item{nclasses}{a numeric value representing the number of unique attribute profiles}
}
\description{
Calculates the attribute profile probabilities for all students
}

