% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetClassProbsFromMus}
\alias{GetClassProbsFromMus}
\title{Obtain attribute profile probabilities from latent variable means}
\usage{
GetClassProbsFromMus(mus)
}
\arguments{
\item{mus}{a 1 by nclasses numeric vector of latent variable means}
}
\value{
a numeric vector of all attribute profile probability values
}
\description{
Gets all attribute profile probabilities given mu parameters (latent variable means)
}

