% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetGammaName}
\alias{GetGammaName}
\title{Obtain name of a particular gamma parameter}
\usage{
GetGammaName(attributes)
}
\arguments{
\item{attributes}{a string of the attributes associated with this gamma parameter}
}
\value{
a string of the form g_effect_attribtues.
}
\description{
Produces a string of the form: g_effect_attribtues representing gamma parameters
}

