% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetGammaNames}
\alias{GetGammaNames}
\title{Obtain names of all gamma parameters}
\usage{
GetGammaNames(required.attributes)
}
\arguments{
\item{required.attributes}{a numeric vector of attributes required to have mastered to get item correct}
}
\value{
a String vector of names for gamma parameters
}
\description{
Gets names of all gamma parameters
}

