% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetLambdaName}
\alias{GetLambdaName}
\title{Obtain name of a particular lambda parameter}
\usage{
GetLambdaName(item, attributes)
}
\arguments{
\item{item}{a numeric value representing item number}

\item{attributes}{a String of the attributes associated with this gamma}
}
\value{
a String of the form: l_item_effect_attribtues.
}
\description{
Gets the name of a particular lambda parameter and produces a string of the
form: l_item_effect_attribtues.
}

