% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetLambdaNamesForItem}
\alias{GetLambdaNamesForItem}
\title{Gets lambda names for a given item}
\usage{
GetLambdaNamesForItem(item.index.intercept, item.index.main.effect,
  required.attributes)
}
\arguments{
\item{item.index.intercept}{a numeric value representing item number for the intercept lambda}

\item{item.index.main.effect}{a numeric value representing item number for the main effects lambda}

\item{required.attributes}{a numeric vector of attributes required to have mastered to get item correct}
}
\value{
a String vector of lambda names for one item
}
\description{
Gets lambda names for a given item
}

