% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetMusFromGammas}
\alias{GetMusFromGammas}
\title{Calcualte attribute means from gamma parameters}
\usage{
GetMusFromGammas(estimates0, pmatrix)
}
\arguments{
\item{estimates0}{a numeric vector of parameter estimates in the form of lambdas and gammas}

\item{pmatrix}{a numeric nclasses by nattributes matrix of all possible attribute profiles}
}
\description{
Give gamma parameters and all attribute profiles, calculates the attribute means (i.e., mus)
}

