% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetParameterResultsMCMC}
\alias{GetParameterResultsMCMC}
\title{Calculate results for attributes and attribute profiles}
\usage{
GetParameterResultsMCMC(parameter.result)
}
\arguments{
\item{parameter.result}{a list parameter estimates for each iterations and chains}
}
\description{
Calculates results for attributes and attribute profiles for all students based on MCMC output
}

