% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetProbCorrectNcrum}
\alias{GetProbCorrectNcrum}
\title{Calculate item probability of correct response for NCRUM parameterization}
\usage{
GetProbCorrectNcrum(pi.i.star.prime, r.ia.star, q.ia, alphas)
}
\arguments{
\item{pi.i.star.prime}{a numeric value for the probability of a correct response for item i given
that no attributes have been mastered.}

\item{r.ia.star}{a numeric 1 x nattributes vector of the penalty probabilities for this item of a correct response
for each required attribute that has not been mastered.}

\item{q.ia}{a numeric 1 x nattributes vector of the qmatrix entries for the given item.}

\item{alphas}{a numeric 1 x nattributes vector with a 1 if a student has mastered an attribute and 0 otherwise.}
}
\value{
the numeric value representing the probability of a correct response for the ith item.
}
\description{
Calculates the probability of a correct response for an item given an NCRUM parameterization
}

