% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetRequiredAttributes}
\alias{GetRequiredAttributes}
\title{Generate required attributes}
\usage{
GetRequiredAttributes(qmatrix, pmatrix)
}
\arguments{
\item{qmatrix}{a nitems by nattributes matrix that specifies which items are required for mastery of each attribute (i.e., latent variable)}

\item{pmatrix}{a numeric nclasses by nattributes matrix of all possible attribute profiles}
}
\value{
a nclasses by nitems character matrix with Strings representing the required attributes for a given item and class.
}
\description{
Generates a matrix of required attributes
}

