% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetRequiredAttributesLambdaNido}
\alias{GetRequiredAttributesLambdaNido}
\title{Calculate required attributes for lambdas for a NIDO model.}
\usage{
GetRequiredAttributesLambdaNido(required.attributes.tau)
}
\arguments{
\item{required.attributes.tau}{a nclasses by nitems matrix with each cell containing a String
representing the attributes required for the given class and item specific to tau parameters.}
}
\value{
a nclasses by nitems matrix with each cell containing a String representing the
attributes required for the given class and item specific to lambda parameters for the NIDO model.
}
\description{
Calculates required attributes for a model parameterized in the LCDM framework using the NIDO
model
}

