% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetThresholdLabels}
\alias{GetThresholdLabels}
\title{Obtain item threshold labels}
\usage{
GetThresholdLabels(qmatrix, pmatrix)
}
\arguments{
\item{qmatrix}{a numeric nitems by nattributes matrix that specifies which items are required for mastery of each attribute (i.e., latent variable)}

\item{pmatrix}{a numeric nclasses by nattributes matrix of all possible attribute profiles}
}
\value{
an nclasses by nitems character matrix with appropriate threshold labels.
}
\description{
Gets item threshold labels for a given qmatrix and pmatrix
}

