% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetThresholdValues}
\alias{GetThresholdValues}
\title{GetThresholdValues}
\usage{
GetThresholdValues(threshold.labels, taus)
}
\arguments{
\item{threshold.labels}{an nclasses by nitems character matrix with appropriate threshold labels.}

\item{taus}{numeric vector of taus correctly named
threshold.labels.cls: String vector of names of thresholds for given class}
}
\description{
GetThresholdValues
}

