% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetThresholdValuesKernelPiR}
\alias{GetThresholdValuesKernelPiR}
\title{Calculate item threshold values for NCRUM parameterization}
\usage{
GetThresholdValuesKernelPiR(estimates0, qmatrix, pmatrix, threshold.labels)
}
\arguments{
\item{estimates0}{a numeric vector of parameter estimates in the form of pis and rs}

\item{qmatrix}{a numeric nitems by nattributes matrix that specifies which items are required for mastery of each attribute (i.e., latent variable)}

\item{pmatrix}{a numeric nclasses by nattributes matrix of all possible attribute profiles}

\item{threshold.labels}{an nclasses by nitems character matrix with appropriate item threshold labels}
}
\value{
a numeric vector of item threshold values
}
\description{
Calculates all item threshold values for an NCRUM parameterization given parameter estimates,
qmatrix, attribute profiles, and threshold label matrix.
}

