% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{head,attribute.class-method}
\alias{head,attribute.class-method}
\title{headattribute.class}
\usage{
\S4method{head}{attribute.class}(x)
}
\arguments{
\item{x}{\code{\link{attribute.class}} input object}
}
\description{
headattribute.class
}

