% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{head,attribute.profile.class-method}
\alias{head,attribute.profile.class-method}
\title{head attribute.profile.class}
\usage{
\S4method{head}{attribute.profile.class}(x)
}
\arguments{
\item{x}{\code{\link{attribute.profile.class}} input object}
}
\description{
head attribute.profile.class
}

