% Generated by roxygen2 (4.0.1): do not edit by hand
\name{InitializeParameters}
\alias{InitializeParameters}
\title{Initialize parameter estimates}
\usage{
InitializeParameters(estimates, nclasses)
}
\arguments{
\item{estimates}{a vector of parameter estimates}

\item{nclasses}{a numeric value representing the number of unique attribute profiles}
}
\description{
Obtains appropriate parameter estimates for a non-kernel model (e.g., taus and nus)
}

