% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LongFormatResults}
\alias{LongFormatResults}
\title{Tranform dataframe to long format}
\usage{
LongFormatResults(d, result.type = "class")
}
\arguments{
\item{d}{a list of MCMC output either classes or parameter estimates}

\item{result.type}{optional character strings containing result.type: \code{class} or \code{parameter}}
}
\description{
Given a dataframe in wide format, transform to long format
}

