% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{observations.test}
\alias{observations.test}
\title{Obervations}
\format{a data frame containing 15 students and 11 items. 1 student per row}
\usage{
observations.test
}
\description{
This data set list the dichotomous responses of students for a test containing 11 items.
}
\author{
Margi Dubal \email{margidubal@gmail.com} & Diane Losardo \email{dlosardo@amplify.com}
}

