% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{parameter.acov.DCM.Mplus.interaction.test}
\alias{parameter.acov.DCM.Mplus.interaction.test}
\title{Covariance matrix of parameter estimates calibrated using Mplus for fully saturated model and Q-matrix with interaction.}
\format{A data frame}
\usage{
parameter.acov.DCM.Mplus.interaction.test
}
\description{
matrix of covariances of all model parameters. If \code{NULL} (the default) model parameters are not randomized
for each iteration of MCMC
}
\author{
Margi Dubal \email{margidubal@gmail.com} & Diane Losardo \email{dlosardo@amplify.com}
}

