% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{parameter.class-class}
\alias{parameter.class}
\alias{parameter.class-class}
\title{parameter.class}
\description{
parameter.class
}
\section{Slots}{

\describe{
\item{\code{results}}{A data frame of parameter estimates}
}}

