% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{parameter.means.DCM.kernel.Mplus.test}
\alias{parameter.means.DCM.kernel.Mplus.test}
\title{Kernel parameter estimates calibrated using Mplus for fully saturated model.}
\format{a vector or a dataframe}
\usage{
parameter.means.DCM.kernel.Mplus.test
}
\description{
a numerical vector of calibrated item and structural kernel parameters. Values must be in the order of \code{\link{GetParameterNames}} if
parametrization method is Mplus and non-kernel parameters are used. If kernel parameters values are used must be in order of \code{\link{GetKernelParameterNames}}
}
\author{
Margi Dubal \email{margidubal@gmail.com} & Diane Losardo \email{dlosardo@amplify.com}
}

