% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{parameter.means.names.DCM.Mplus.test}
\alias{parameter.means.names.DCM.Mplus.test}
\title{Names of parameter estimates calibrated using Mplus for fully saturated model and Q-matrix with  no interaction.}
\format{A vector of string}
\usage{
parameter.means.names.DCM.Mplus.test
}
\description{
a vector of string containing names of calibrated parameter estimates.
}
\author{
Margi Dubal \email{margidubal@gmail.com} & Diane Losardo \email{dlosardo@amplify.com}
}

