% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{print,attribute.class-method}
\alias{print,attribute.class-method}
\title{print attribute.class}
\usage{
\S4method{print}{attribute.class}(x)
}
\arguments{
\item{x}{\code{\link{attribute.class}} input object}
}
\description{
print attribute.class
}

