% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{print,attribute.profile.class-method}
\alias{print,attribute.profile.class-method}
\title{print attribute.profile.class}
\usage{
\S4method{print}{attribute.profile.class}(x)
}
\arguments{
\item{x}{\code{\link{attribute.profile.class}} input object}
}
\description{
print attribute.profile.class
}

