% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{qmatrix.test}
\alias{qmatrix.test}
\title{Q-matrix}
\format{a matrix containing 11 items and 3 attributes}
\usage{
qmatrix.test
}
\description{
This data set list 1s and 0s indicating relation between items and attributes. This matrix specifies which items are required
 for mastery of each attribute (i.e., latent variable). The matrix must be a size of \code{nitems X nattributes}
}
\author{
Margi Dubal \email{margidubal@gmail.com} & Diane Losardo \email{dlosardo@amplify.com}
}

