% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SampleParameterEstimates}
\alias{SampleParameterEstimates}
\title{Randomly sample parameter estimates}
\usage{
SampleParameterEstimates(means, acov)
}
\arguments{
\item{means}{a numerical vector of calibrated item and structural parameters}

\item{acov}{a numerical matrix of covariances of item and structural parameters}
}
\value{
A vector of randomly sampled parameter estimate values
}
\description{
Given parameter estimates and asymptotic covaraince matrix of parameter estimates, randomly samples a new
set of estimates
}

