% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{summary,attribute.profile.class-method}
\alias{summary,attribute.profile.class-method}
\title{summary attribute.profile.class}
\usage{
\S4method{summary}{attribute.profile.class}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{\code{\link{attribute.profile.class}} input object}

\item{verbose}{a logical. If TRUE, additional diagnostics are printed.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
summary attribute.profile.class
}

