% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{summary,dcm.scorer.class-method}
\alias{summary,dcm.scorer.class-method}
\title{summary of dcm.scorer.class}
\usage{
\S4method{summary}{dcm.scorer.class}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{\code{\link{dcm.scorer.class}} input object}

\item{verbose}{a logical. If TRUE, additional diagnostics are printed}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
summary of dcm.scorer.class
}

